requirements.txt
customtkinter==5.2.2
pandas==2.2.3
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
future~=1.0.0
tkintermapview~=1.29